Comment機能に関するファイルをfeatures/articles/_commentsに移動
