--- frontend/app/app/components/atoms/tags/postComment/theme/styleMap.ts
@@ -1,7 +1,0 @@
-import type { MapTagColors } from "./type";
-import type { candidateTagData } from "@/app/constants/datas/portfolios/comments/tags";
-
-export const postCommentTagStyles: MapTagColors<typeof candidateTagData> = {
-  "bug report": { label: "bug report", colorTheme: "red" },
-  suggestion: { label: "suggestion", colorTheme: "blue" }
-};
