--- frontend/app/app/components/molecules/posts/comments/PostCommentCardFooter.tsx
@@ -1,27 +1,0 @@
-"use client";
-
-import { CardFooter, Divider, Flex, Stack } from "@chakra-ui/react";
-
-import { useCommentData } from "@/app/hooks/datas/useCommentData";
-import UpdatedDateText from "@/app/components/atoms/texts/UpdatedDateText";
-
-import PostCommentTags from "./PostCommentTags";
-
-const PostCommentCardFooter: React.FC = () => {
-  const commentData = useCommentData();
-  const { updatedAt, tags } = commentData;
-
-  return (
-    <CardFooter>
-      <Stack w="full">
-        <Divider />
-        <Flex justifyContent="space-between" direction="row-reverse">
-          <UpdatedDateText date={updatedAt} />
-          <PostCommentTags tags={tags} />
-        </Flex>
-      </Stack>
-    </CardFooter>
-  );
-};
-
-export default PostCommentCardFooter;
