--- frontend/app/app/components/organisms/posts/comments/PostCommentCard.tsx
@@ -1,27 +1,0 @@
-"use client";
-
-import { Card, Spacer } from "@chakra-ui/react";
-
-import { useCommentData } from "@/app/hooks/datas/useCommentData";
-import { useGetLoginState } from "@/app/hooks/recoil/loginState/useGetLoginState";
-import PostCommentCardBody from "@/app/components/molecules/posts/comments/PostCommentCardBody";
-import PostCommentCardFooter from "@/app/components/molecules/posts/comments/PostCommentCardFooter";
-import PostCommentCardHeader from "@/app/components/molecules/posts/comments/PostCommentCardHeader";
-
-const PostCommentCard: React.FC = () => {
-  const { isLogin, userData } = useGetLoginState();
-  const commentData = useCommentData();
-  const isUserComment = isLogin ? commentData.user.id == userData.id : false;
-
-  return (
-    <Card w={{ base: "auto", md: "md" }} rounded="none">
-      <PostCommentCardHeader isUserComment={isUserComment} />
-      <Spacer my="-3" />
-      <PostCommentCardBody />
-      <Spacer my="-3" />
-      <PostCommentCardFooter />
-    </Card>
-  );
-};
-
-export default PostCommentCard;
