+++ frontend/app/app/features/articles/_comments/components/layouts/ArticleCommentCard/header/index.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import { CardHeader, Flex, Heading } from "@chakra-ui/react";
+
+import UserIcon from "@/app/components/atoms/users/UserIcon";
+
+import OptionCommentMenuButton from "./sections/OptionCommentMenuButton";
+import { useCommentData } from "../../../../hooks/useCommentData";
+
+interface Props {
+  isUserComment?: boolean;
+}
+
+const ArticleCommentCardHeader: React.FC<Props> = ({ isUserComment = false }) => {
+  const commentData = useCommentData();
+  const { user } = commentData;
+  const image = user.image || "/defaultUserImage.png";
+
+  return (
+    <CardHeader>
+      <Flex>
+        <Flex flex="1" gap="2" alignItems="center" flexWrap="wrap">
+          <UserIcon image={image} name={user.name} diameter={25} />
+          <Heading fontSize="sm">{user.name}</Heading>
+        </Flex>
+        {isUserComment && <OptionCommentMenuButton />}
+      </Flex>
+    </CardHeader>
+  );
+};
+
+export default ArticleCommentCardHeader;
