+++ frontend/app/app/features/articles/_comments/components/layouts/ArticleCommentCard/header/sections/OptionCommentMenuButton.tsx
@@ -1,0 +1,26 @@
+"use client";
+
+import { IconButton, Menu, MenuButton, MenuList } from "@chakra-ui/react";
+import { AiOutlineMore } from "react-icons/ai";
+
+import DeleteCommentButtonWithModal from "../../../../DeleteCommentButtonWithModal";
+import PatchCommentButtonWithModal from "../../../../UpdateCommentButtonWithFormModal";
+
+const OptionCommentMenuButton: React.FC = () => {
+  return (
+    <Menu>
+      <MenuButton
+        as={IconButton}
+        aria-label="Options"
+        icon={<AiOutlineMore />}
+        variant="ghost"
+      />
+      <MenuList>
+        <PatchCommentButtonWithModal />
+        <DeleteCommentButtonWithModal />
+      </MenuList>
+    </Menu>
+  );
+};
+
+export default OptionCommentMenuButton;
