+++ frontend/app/app/features/articles/_comments/types/api.ts
@@ -1,0 +1,8 @@
+export interface CommentParams {
+  content: string;
+  tagIds?: number[];
+}
+
+export interface CommentFormParams extends Omit<CommentParams, "tagIds"> {
+  tagIds: string[];
+}
