+++ frontend/app/app/features/articles/components/layouts/ArticleCard/footer/index.tsx
@@ -2,9 +2,9 @@
 
 import { CardFooter, Divider, Flex, Stack, Text } from "@chakra-ui/react";
 
+import CreateCommentButtonWithFormModal from "@/app/features/articles/_comments/components/CreateCommentButtonWithFormModal";
 import LikeButton from "@/app/features/articles/_likes/components/layouts/LikeButton";
 import { useToggleLikeArticleGood } from "@/app/features/articles/_likes/hooks/useToggleArticleGood";
-import CreateCommentButtonWithModal from "@/app/components/organisms/apiActionButtons/posts/comments/CreateCommentButtonWithModal";
 
 import type { ArticleData } from "@/app/features/articles/types/articleData";
 
@@ -23,7 +23,7 @@ const ArticleCardFooter: React.FC<Props> = ({ articleData }) => {
         <Flex justifyContent="space-between" alignItems="center">
           <Flex alignItems="center">
             <LikeButton onClick={toggleLike} isLiked={isLiked} totalLiked={totalLiked} />
-            <CreateCommentButtonWithModal articleData={articleData} />
+            <CreateCommentButtonWithFormModal articleData={articleData} />
           </Flex>
           <Text fontSize="sm" color="blackAlpha.500" my="auto">
             {updatedAt.toLocaleDateString()}
