+++ frontend/app/app/features/articles/types/articleData.ts
@@ -1,6 +1,6 @@
+import type { CommentData } from "../_comments/types";
 import type { candidateOperationStatusData } from "@/app/constants/datas/portfolios/operationStatuses";
 import type { UserData } from "@/app/types/auth";
-import type { CommentData } from "@/app/types/axios/portfolio/comment/comment";
 
 interface Good {
   userId: number;
