--- frontend/app/app/libs/axios/portfolio/comment/patchPortfoliosByIdCommentsById.ts
@@ -1,17 +1,0 @@
-import addAuthInfoToRequest from "@/app/libs/cookie/loadAuthInfo";
-
-import api from "../../api";
-
-import type { PostCommentParams } from "@/app/types/axios/portfolio/comment/comment";
-
-export const patchPortfoliosByIdComments = async (
-  portfolioId: number,
-  commentId: number,
-  params: PostCommentParams
-): Promise<void> => {
-  await api.patch(
-    `/articles/${portfolioId}/comments/${commentId}`,
-    { comment: params },
-    addAuthInfoToRequest({})
-  );
-};
